study-project/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
firebase-admin==6.4.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiofiles==23.2.1
psutil==5.9.6
requests==2.31.0 